feat: redesign now playing card. remove top tracks section
