+++ src/app/now/music.tsx
@@ -3,17 +3,15 @@ import cx from 'classix';
 import { Heading } from '@/components/core/heading';
 import { NowPlaying } from '@/components/views/dashboard/now-playing/now-playing';
 
-import { TopSongs } from './top-songs';
-
 export const MusicData = () => {
   return (
     <article id={'music'} className={cx('flex flex-col gap-16')}>
       <Heading $as={'h2'} className={cx('text-xl')}>
         Music
       </Heading>
-      <div className={cx('flex flex-col rounded-8 border border-divider')}>
+      <div className={cx('grid grid-cols-1 gap-12 mobile-md:grid-cols-2')}>
         <NowPlaying />
-        <TopSongs />
+        <NowPlaying />
       </div>
     </article>
   );
